refactor: separate TiDB walk-through code and improve consistency
- Move TiDB-specific walk-through code to walk_through_for_tidb.go
- Rename walk_through_for_pgantlr.go to walk_through_for_pg.go for consistency
- Rename pgAntlrWalkThrough to pgWalkThrough for consistent naming
- Move PostgreSQL type assertion from walk_through.go to pgWalkThrough
- Remove unused imports from walk_through.go
- Standardize all engine walk-through functions to accept `ast any` parameter
- Simplify WalkThrough dispatcher by consolidating type assertions

All three engine implementations now follow the same pattern:
- mysqlWalkThrough in walk_through_for_mysql.go
- pgWalkThrough in walk_through_for_pg.go
- tidbWalkThrough in walk_through_for_tidb.go
